+++ src/main/java/woowacourse/auth/ui/dto/LoginCustomer.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.ui;
+package woowacourse.auth.ui.dto;
 
 import javax.validation.constraints.Positive;
